requirements.txt
# Please see the INSTRUCTIONS pdf to learn how to use this application

pyglet~=1.5.15
numpy~=1.20.1
Pillow~=8.1.2
